HYD-6215 Unit tests framework script not using defaults
We have a script that sets defaults for our testing in that was developed
to DRY our test setup some. However, the unit tests script was never
updated to use this, and still is duplicating some of its behavior.

We should use the common setup and remove any overlapping setup code.

  - change tests/framework/unit/jenkins_steps/main to source
    tests/framework/utils/defaults.sh in place of its duplicative
    setup.

Change-Id: I5517ecf3871b387013acfca18ab3051bd832f565
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21470
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21662