HYD-6190 Error removing server: Cannot read property 'length' of undefined
When removing servers the host information removed can contain locks as an
empty dictionary.

"locks": {},

However locks is expected to contain a list of read and write locks so an
empty list would be

locks = {'read': [],
         'write': []}

The root cause is that locks is initialized to an empty dictionary rather than
a pair of empty read/write list of locks.

The empty list is possible when a request is made for an object that is deleted
during the read process, i.e. the API end point succesfully fetches the object
but before it fetches the locks the object is deleted.

The simple fix is to make the default returns for locks

locks = {'read': [],
         'write': []}

 - Add correct default locks to JobScheduler.get_locks

Change-Id: I8ab3dc987137b0af39e084604433a8eda13da9d6
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21471
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>