HYD-6108 Test Host Validation
We require tests to check the validation methods used when
adding a server. This ticket implements three of those tests.

The first test is to check that when trying to add a server,
with a non-existent host name it fails as expected.

The second test is to check if the authentication type
is ’private_key_choice’, it fails with a invalid key and works
with a valid private key.

The third test is to check if the authentication type is
'id_password_root’, it fails with an invalid root password
and works with a valid password.

Two functions was added to the RealRemoteOperation class,
the first was to be able to rename a file on a remote address and
the second was to be able to read a file and return its contents from
a remote address.

   - Run the add test server command with a non-existent host name
   - Run the add test server command with invalid and valid private key
   - Run the add test server command with invalid and valid root password
   - Ensure that no functionality works for this command, i.e. no ping, no resolve, etc.
   - Add function to allow rename of file on remote address
   - Add function to allow read and return of file content

Change-Id: I377eb64c558daa6895437a72abcd8fca3e7be39a
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/21473
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>