HYD-6220 Make a central place to control skipping tests
selective_auto_pass is currently only included in the test we
want to skip for gui bumps, and then it skips if its a gui bump.
We'd like to use it as a central place for managing if tests are
supposed to be skipped or not, so we can also check if we should
skip based on the distro.

To be a central place for all suties, it needs to be included in
all of the test suites, and then needs some logic to decide if it
should skip or pass when invoked. We can then check there if it is
a supported distro on teh branch currently so we only run tests on
the supported distro(s).

  - source selective_auto_pass in service and unit tests too
  - make selective_auto_pass skip on gui bump only if it is not
    in a required list of tests that need to pass for gui bumps.
  - skip the tests in selective_auto_pass if it is not a supported
    distro.

Change-Id: I346a4ec2513671fc5a058e8fab770564d5e333ac
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21487
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>