HYD-6262 To work effectively ZFS needs to list either zpools or datasets but not both
Because of an unexpected way in which the hierarchy of volumes and devices
works in IML we need to provide either the zpool volume or the dataset volume
to the user but not both. If we provide both then whilst a dataset volume will
be identified as unavailable if it is in use as a lustre target, the zpool will
still be shown as available.

To resolve this a zpool and dataset (or zvols) should be exclusive. The agent
should report zpools with not children or just the children. The manager must
then delete the children or parent as appropriate.

This implementation will have limited scope of change and should not require 
any additional changes or knock on effects within IML.

This patch updates the agent to return either a zpool or datasets and/or zvols
the children of the zpool/dataset(s)/zvols(s) will be the volumes used to create
the dataset. If a zpool then those volumes will have a single parent but if multiple
datasets or zvols or a mixture of both exist then the volumes will have multiple
parents. This graph rather than tree structure is something already supported by IML.

 - Modify zfs.py to treat dataset/zvols/zpools as peers rather than as a hierarchy
 - Fixup device_helper to realise that directories or files may not exist.
 - Update tests appropriately.

Change-Id: I17bc3060126bead8d270e6c2f4825f6977ef19b8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21697
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>