HYD-6340 Linux devices should disappear when they are deleted
Historically with IML when a device such as a drive is removed the agent or
manager must be restarted for the device to be removed from IML. However with
ZFS devices become much more dynamic and so can appear and disappear often.

HYD-6261 allow device resources to be more easily enumerated and so the
linux.py plugin can be expanded to delete at leave some of the volumes that
appear and disappear. Fortunately the devices that are easy to apply this to
are the 'special devices' such as lvm and zfs.

Add a feature that upon update of the volumes list scan the existing list and
removes anything that has not been reported. This change ensures the volumes
are deleted ones all volume_nodes (LinuxDeviceNodes) have been removed for that
device.

 - Detect missing devices in linux.py special cases and remove them from
   resources
 - When removing from resources remove the drive resource if all
   LinuxDeviceNodes have disappeared

Change-Id: If7ad9c11f8827a1daa8e7bb64f2697f8e132549b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21921
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>