HYD-6266 Bump nginx to 1.10.1-1
We currently are shipping nginx-1.8.1-1. This version has support for SPDY,
which allows for faster delivery of web content.

Unfortunately, SPDY support has been removed
from Chrome as of May 15th 2016.

Nginx has added support for the HTTP2 protocol
(which is based somewhat on SPDY) in version 1.9.5.
and is supported by most browsers.

In 3.1, multiplexing / avoiding connection limit blocking will be
more important than in the past due to our approach of
loading some pages in iframes, which will contain
~ 100 new requests when the page loads.

  - Bump Nginx to 1.10.1-1.

Change-Id: I9a6ee3dfc1ef9bd7783741d5d3c8dd5d80cd3b52
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/21716
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>