HYD-6030 Improve SSI test functionality for ZFS and LDISKFS
Additional changes to make ZFS testing work. These patch does not yet introduce
the actual changes to test on ZFS but creates some additional functionality
to make clearing and creation of ZFS pools for testing more versatile and
robust.

A series of bitwise flags are added to the clear_zfs_pools function which allows
zpools and datasets to be cleared and created with some granularity. A flag is
added to the json to allow zpools to be recreated every test or only on the first
test.

Make TestConfigureZfsTargets override quick_setup to force it to be true for
that case. This is required because because at the time of this call no-zfs
pools exist and so the cleanup routine in the regular setup will fail.

 - Make TestConfigureZfsTargets override quick_setup to force
 - Add CZP_* flags to cleanup_zfs_pools to allow granularity of functionality
 - Simplify TestConfigureZfsTargets to use new granularity

Change-Id: Ib27f7e4bae120fe7010243856a08c6537bce600c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/23173
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>