HYD-6030 Run SSI tests with ZFS and LDISKFS
Activate zfs devices in json and fixup an issue Chris introduced when he
removed quick_setup from the TestConfigureZfsTargets tests.

Run the SSI test suite with ZFS-backed lustre targets in addition to LDISKFS.
This gives us some assurance with regard to the functionality of managed ZFS.

The solution is to provide a new entry in the json which stipulates if a disk
is zfs or ldiskfs. As the first test in a file called test_aaa_zfs_setup.py
we then convert the linux device to zpools and update the configuration
appropriately.  The rest of the code then correctly, silently and without its
knowledge uses zpools rather than raw block device.

This patch also sets a flag to cause zpools to be recreate for each test
this flag covers up corruption of zpools and in real live user data
that can occur in test.

 - SSI json updated to have 3 zfs devices and 2 raw-devices (ldiskfs)
 - Set the flag to cause zpools to be recreated every test

Change-Id: I6c0338bb2eeef9f9f163e248906cda813dd0588d
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22647
Tested-by: Jenkins <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User