HYD-6271 Reduce linux scan times by not retrying failing paths
HYD-5826 increased scan times for Linux devices due to retry of
os.stat calls to catch temporarily disappearing devices.

This patch does not retry a path that has failed previous initial
retries.

So if it tries to look at

/dev/ssss

and after however many retries it still fails, the path is added
to a set. Then next time the path is accessed, it should try once
and fail. If a path succeeds it should be removed from the set.

  - make relevant changes to device_helper
  - update unit tests

Change-Id: I1d5bd377f8f0aad17356f381ac7d3c434ec37c52
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/21856
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>