HYD-6105 Test anonymous access control
Automated test which checks that the
ALLOW_ANONYMOUS_READ setting works to allow/disallow
anonymous access to the system.

This was achieved by verifying anonymous access is granted to
the api endpoints of the manager when the settings allow anonymous
access to the system. Then changing these settings by creating a
local_settings.py file which is read by settings.py changing
‘ALLOW_ANONYMOUS_READ’ to false. Restart chroma-config
to have the new settings active and retry accessing the api endpoints,
this time verifying that it is unauthorised to the api endpoints. Whilst also
checking access is granted when the request has the correct login credentials.

    - Add function to remote_operation class to create file on remote host
    - Add function to remote_operation class to delete file on remote host
    - Add function to ApiTestCaseWithTestReset class to allow
      chroma-manager http request without any authentication credentials
    - Ensure access is granted when the ‘ALLOW_ANONYMOUS_READ’ is set to True
    - Ensure access isn’t granted when the ‘ALLOW_ANONYMOUS_READ’ is set to False
    - Ensure access is granted with the right login credentials when the
      'ALLOW_ANONYMOUS_READ’ setting is set to False

Change-Id: Id345217b69cf4b627de1ea37c568e1eba2ba39db
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/21912
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>