HYD-6341 Simulator: cluster-sim cli broken on develop
cluster-sim register --username admin --password lustre
does not work on the tip of develop.

This is because the the name of the base_managed profile has changed to
base_managed_rh7.

When invoking from the cli, the base_managed profile name is still used.
Change the reference of base_managed to base_managed_rh7.

Change-Id: I91e4c0e68dce995156fd6ad0d98cced194156e56
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/21931
Tested-by: Jenkins
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>