HYD-6325 Remove gui/selenium tests
The selenium/gui tests were decommissioned months ago in EE 3.0.0.0.
However, the test aggregation was never updated to not require it,
so Jenkins continues to have the chroma-selenium-tests job run to
spit out the fake pass xml. We should clean out the old dead suite.

This script removes chroma-selenium-tests as a required test suite
and the dead scripts that were used to run them.

  - delete chroma-manager/tests/framework/gui/
  - remove references to chroma-selenium-tests from
    chroma-manager/tests/framework/test_reporting/jenkin_steps/main

Change-Id: I5faa744045be69c6c7ee4e43d9c4d04feb503454
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21957
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/22076