HYD-6353 Remove hard code of ieel tarball name in test framework
Our test scripts have the 'ieel' in the tarball name hardcoded all
over the place. This makes updating the tarball name a pain as one
has to find all of the references to it and update.

It would be better to obtain this information from
the Makefile as we do the ieel_version so that when you change the
tarball name you don't have to update all these scripts too.

  - Add a .short_archive_name to Makefile.version to expose the
    variable used to set this in building to outside scripts,
    like we already do for .ieel_version.
  - Add SHORT_ARCHIVE_NAME and ARCHIVE_NAME to defaults.sh so all
    test framework scripts have access to this information.
  - Update the hardcoded references to 'ieel' in
    chroma-manager/tests/framework/ to use the variables instead.

Change-Id: Id271def5429579f90ae0651252b979e117e7ef09
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21958
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/22078