HYD-6362 SSI tests use a conf param unavailable on a ZFS target
Conf param tests use writethrough_cache_enable as a test parameter but this is
unavailable on ZFS targets. To make the test universal we should use use a
parameter that is available on all backend filesystems. sync_journal
would seem like a good choice.

A shortcoming of the current test is that only 1 value is set and so we cannot be
sure that the value would not have occurred anyway. Whilst not a complete fix for
all situations we can add 2 OSTs and set different values for each OST which
asserts that different actions have different outcomes.

A bug in the simulator where the odbfilter values always reported the default
value required fixing because the ost conf param value is now not tested by
setting it to the default values.

Add sync_journal to the IML conf parameters so that we can use them for
test.

Fix fake_server to store and retrieve obdfilter conf parameters

 - Add sync_journal to conf_param.py
 - Change tests to use sync_journal in place of writethrough_cache_enable
 - Change tests to use 2 osts and write different sync_journal values to
   each OST so that we can be sure different actions have different outcomes
 - Update fake_server to work with sync_journal
 - Fix fake_server to store and retrieve obdfilter conf parameters

Change-Id: Idf7526d54948c52246caab398b6d4dd8f685f901
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22191
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>