HYD-6358 Reload of old GUI on Error does not reload top window
The old GUI does not take into account that the app now uses frames. A
side effect of this is that if the old error modal appears and the
"reload" button is clicked, it reloads a new instance of the app in the
current frame, giving the impression that there are two tree views.

To solve this problem the reload link needs to be updated to reload the
top target (window) as opposed to the current frame. This patch will
update the link on the error modal to do just that.

Change-Id: I8c722ad349a9d5f02a0e1c158d25e75f4fe568d4
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/21997
Tested-by: Jenkins
Reviewed-by: Paul Kline <paul.everett.kline@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>