HYD-6364 TestFilesystemSameName_HYD832 fails on zfs because datasets exist
When a filesystem is deleted we do not by design remove the datasets. This is
so that the filesystem can be rebuilt in the event of an accident. We leave it,
today, to the user to remove the datasets.

However we do not allow a filesystem to be built on a zpool with a dataset and
so for this reason the TestFilesystemSameName_HYD832 test fails because it
creates, removes and recreates a filesystem.

The solution is to remove the datasets in the test so that a filesystem can be
created.

 - Modify test to call cleanup_zfs_pools with a list of the zpool that are
   going to be reused. This causes the datasets to be removed and allows them
   to be reused.

Change-Id: Ia93abd34e8e5abdda5da07649e1f5a75625fd8ca
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22359
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>