HYD-6379 Volumes without associated VolumeNodes don't get marked as deleted
When all VolumeNodes are dis-associated from a Volume, it becomes
"orphaned" and is not marked as deleted when we remove hosts. This
can happen when we format a target and the VolumeNode is switched to
that referring to the dataset rather than the zpool. The Volume that
refers to the zpool can then end up "orphaned".

This issue exhibits itself during 'graceful_shutdown' in tests when
checking if all Volumes have been cleared from the database.

A work around to this issue is to remove the orphaned volumes durring
accesses to the volumes end point, this is a workaround with the
correct solution being to prevent orphans being created in the first
place. Time prevents us resolving the issue correctly at this point
in the development process. A side effect of the workaround is additional
database queries are created durring accesses to the volumes endpoint
however because the volumes endpoint can and should be accessed using
long-polling these additional queries are an acceptable side effect

  - When applying filters for a VolumeResource, mark volume as
    deleted if no associated VolumeNodes exist
  - Update unittests to allow for additional queries on the volumes
    endpoint.

Change-Id: Ibdb12d0e3d10913da0e5d9e3efb67ce171e2c865
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22105
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>