HYD-6396 Volume API shows VolumeNodes that are deleted
If a volume was present on VolumeNode with id = XXXX, then a query to the
volumes end point with host_id=XXXX will return that volume. However the code
does not check that the VolumeNode has not been deleted and so in the case of
a Volume being removed from a single Node it will still appear on that node.
This is an old bug but one that has been highlighted and made visible by the
tree-view.

Simply checking for volumenode__not_deleted=True makes the code work
correctly.

 - Add checks for volumenode__not_deleted=True to the Volume api
 - Add tests to check that deleted VolumeNodes are not reported

Change-Id: Idb8f38430e61f2c63e75fd894617c006536ffbdb
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22248
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>