HYD-6420 REST interface responds incorrectly to a non existent filesystem id
When requesting endpoints with no-values returned we expect and empty return.
https://lotus-50vm13/api/target/?format=json&id=100

{
meta: {
limit: 20,
next: null,
offset: 0,
previous: null,
total_count: 0
},
objects: [ ]
}
however when selecting a non-existant filesystem id for target, host or volume
we return a 404 error and unique method.
https://lotus-50vm13/api/target/?format=json&filesystem_id=100

{
filesystem: "No ManagedFilesystem matches the given query."
}

These issues cause the treeview to receive errors with a knock on effect that
the tree can become out of step with the actual backend.

The solution is to detect a non-existent filesystem and produce a filter that
returns no objects in the target resource. The approach taken is to filter on
id=-1 which will always produce zero hits.

 - Remove use of get_object_or_404 and replace with an approach that produces
   a queryset returning nothing in the case of the filesystem id not being
   present. Do this for volume, host and target. Subtle differences mean that
   a dry approach did not make sense.
 - Add tests for the three endpoints to validate the behaviour works as expected.

Change-Id: Ieab646dd9dd4eddbcb15427da1a4baffc5702faa
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22288
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>