HYD-6441 Tests failing with VolumeNode.DoesNotExist
The VolumeNode error is due to the delay in creating VolumeNodes
during FormatTargetJob in the MkfsStep. A volume is created for the
new dataset (for a ZFS target) and after a delay, the relevant
VolumeNodes for each host are created by the resource manager. If
these VolumeNodes are not present by the time the
UpdateManagedTargetMount step (still in FormatTargetJob) queries the
volumenode table for entries referencing the dataset path and
relevant hosts, a VolumeNode.DoesNotExist exception is raised.

This exception is seen reliably in tests that create a filesystem
with 2 OSTs (with the SSI-ZFS changes enabled).

The solution involves adding a method that waits for the VolumeNodes
to be created.

  - create method wait_for_success which will monitor any exceptions
    raised during the execution of an anonymous (lambda) function
    for expected exception instances and will fail if the function
    still raising expected exceptions after a given timeout
  - wrap the get() query for VolumeNode within
    UpdateManagedTargetMount in wait_for_success() routine
  - remove UpdateDevices step from FormatTargetJob to avoid
    unnecessary device querying, polling should suffice

Change-Id: I29e7a17e3331787070622922ea7090e20f8d7021
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/22319
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>