HYD-6030 Changes to run SSI tests with ZFS and LDISKFS
This are the changes to run the SSI test suite with ZFS-backed lustre targets
in addition to LDISKFS. But do not actually enable the ZFS tests, this is to
enable patches dependent on these changes to land.

The solution is to provide a new entry in the json which stipulates if a disk
is zfs or ldiskfs. As the first test in a file called test_aaa_zfs_setup.py
we then convert the linux device to zpools and update the configuration
appropriately.  The rest of the code then correctly, silently and without its
knowledge uses zpools rather than raw block device.

TestBlockDevices needed o be updated and so new code to remove and created
datasets as part of HA cleanup added.

  - Remove and recreate zools and edit config within test
    test_aaa_backfs_setup.py
  - _execute_commands and _execute_simultaneous_commands moved to
    UtilityTestCase.execute_commands
  - Precommit hook added to remove E711 and E712 after agreement with
    Tom (not worth a patch on its own)
  - SSI json updated to have 5 raw-devices (ldiskfs)
  - Simulator json updated to have 5 raw-devices (ldiskfs)

Change-Id: Ia6f430db4c552c6be473a15aec538946100815a3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22326
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>