HYD-6381 Fix temporary resolution for issue described in HYD-6379
The temporary solution in HYD-6379 (2c04706ac33dc12574f26a3050184a1eb8e185fa)
was a true fix rather than root cause and flawed because it didn't clean up
the resources associated with the Volume.  A better soution is to make the
resource manage correctly remove Volumes and VolumeNodes when they are deleted
and handle appropriately the storage_resource deletion.

Change-Id: Ic84ef89101174c00331767b0b753295af75d75fe
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22778
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>