HYD-6445 GUI: Firefox doesn't load
This is because we are using a RC4 ssl cipher that has a known weakness:
https://www.imperva.com/docs/hii_attacking_ssl_when_using_rc4.pdf

Instead of using RC4, blacklist it and replace with the following cipher
list:

ECDH+AESGCM:ECDH+AES256:ECDH+AES128:DH+3DES

Also, enable the Strict-Transport-Security for the /ui location.

This will allow Firefox to work again.

  - Update ssl_ciphers.
  - Add Strict-Transport-Security.
  - Add favicon for help pages.

Change-Id: I4ef209fa81ee9b5630c59e2f71377413777de914
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/22337
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>