HYD-6447 Tidy up the test_hosts_add_existing_filesystem test
The motivation for this change is that debugging a failure on this
test which turned out to be unrelated to the test was made difficult
because the test itself was written in a misleading way. It gave the
impression that new volumes were created when the second host
was installed - which is not the case.

Simplifying the test, which was required to understand the actual
cause of the failure in HYD-6447, makes the test easier to understand
and hense the changes are well worth the effort of landing.

Formalise the quick_setup feature the was added as part of HYD-6030
so that it is useful without being dangerous. A new environment
variable called IML_QUICK_TEST_SETUP can be used in development to
speed up test runs.

  - Use wait_usable_volumes rather than get_usable_volumes
  - Ensure new volume is mounted on newly added host
  - Remove unnecessary list comprehension processing usable volumes
  - Minor reformatting
  - Formalise the quick_setup feature the was added as part of HYD-6030
    so that it is useful without being dangerous. A new environment
    variable called IML_QUICK_TEST_SETUP can be used in development to
    speed up test runs.

Change-Id: Idfc7c61d209ea7bb6223e74ea26d94281cbebcc7
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22360
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>