HYD-5037 ZFS simultaneous import protection needs hostid set
Failover on ZFS needs to provide simultaneous import, this support has
been integrated into Lustre and requires spl_hostid to be set for the
spl module. zfs reads hostid out of {{/etc/hostid}} on load.

The solution involves IML using genhostid to assign a random id to the 
/etc/hostid file of a ha node during the managed host setup.

To implement this test required a new method for remote operations
file_exists and when adding this to SimulatorRemoteOperations I realised
that the other operations pass rather than throwing an error. This is
dangerous as it could make sim tests appear to work, all file (recent)
pass's are changed to RuntimeErrors.

Reference: HYD-5037 and LU-7134

  - Add InitialiseBlockDeviceDriversStep to invoke relevant agent routine
  - Add InitialiseBlockDeviceDriversStep to SetupHostJob to run on managed hosts
  - Add initialise_block_device_drivers agent routine in manage_node.py to
    call initialise_driver() on each BlockDevice subclass
  - Add dummy initialise_driver routine to BlockDevice baseclass
  - Add concreted implementation of initialise_driver routine to BlockDeviceZfs
    subclass to check existence of hostid file and generate if absent
  - Add unit and integration test coverage, skip simulator integration test
  - Add initialise_block_device_drivers entry to fake_action_plugins
    and cleanup some fake sim code
  - Add initscripts requirement to agent spec file to ensure genhostid util
    availability
  - Add file_exists to RemoteOperations

Change-Id: Ie1332b834969c22ec80690e85466099b44926493
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/22464
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>