HYD-6104 Storage server authentication
The manager creates a security file for the storage server.
The manager will not accept any communications from the
storage server without it’s contents being correct.

Test that a active alert is raised when you change the contents
of the security file and then change back the security file to the original
content.Then check that the active alert has gone.

    - Add function to remote_operation class to copy file on remote host
    - modifyithe private.pem file and wait for the alert the alert to be raised
    - modifying the private.pem file back to original contents
    - assert that active alert raised has gone

Change-Id: I7752b064371d4bda7cd29336f4041a3175aaa61a
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/22407
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>