HYD-6487 A removed disk/volume requires an agent restart to update
Removing a disk/volume from a server requires the agent to be restarted before
the manage reports the event. With ZFS this is a problem because Volumes in the
form of dataset/zpools can be created and removed very easily.

The resource manager needs to be updated to deal with the removal of devices
buy culling resources whenever a global resource is removed and also be
deleting local resources when scoped resources are removed.

To fully implement this patch the Linux plugin will also need to be updated to
remove resources from its pool when they are update. This will be another
ticket.

 - Update resource_mgr to cull resources when a global resource is removed.
 - Update tests to ensure that this behaviour is correct.

Change-Id: I9124fa56b4d1437b5f611c006e8f3b17c40f5d6b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22331
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>