HYD-6488 The Linux plugin needs to actively report removed resources
Removing a disk/volume from a server requires the agent to be restarted before
the manage reports the event. With ZFS this is a problem because Volumes in the
form of dataset/zpools can be created and removed very easily.

The Linux plugin needs to be updated to deal with the removal of disks/volumes
by removing appropriately from the resource database.

To fully implement this patch the resource_mgr will also need to be updated to
remove resources from its pool when they are update. This will be another
ticket.

 - Update Linux plugin to remove unused resources as they are deletde
 - Update tests to ensure that this behaviour is correct.

Change-Id: I8289ef0cb2e02b784ec740cca32e9dfd30558367
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22556
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>