HYD-6470 mark_deleted commits transactions
The DeletableMetaclass has a mark deleted method that is decorated with
transaction.commit_on_success. This means that when called from within a
transaction the calling transaction is silently committed. This is not useful
behaviour.

mark_deleted required updating to be atomic when not in a transaction and
nothing more than regular code when called from within a transaction.

 - Renamed mark_deleted to _mark_deleted and remove the decorator
 - Add a wrapper called mark_deleted that wraps with commit_on_success if
   the call is not within a transaction.

Change-Id: I503e4b464bcd10a309b8256f33b68680f78c8ec5
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22461
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>