HYD-6606 Support upgrade testing on RHEL7
When we initially introduced support for el7, the upgrade testing
was still testing upgrading from el6, as there was no previous el7
release to test upgrading from. Now on EE 3.1, when we switched the
starting version to EE 3.0 on el7, the test framework scripts were
shown to be missing some updates to make them work on RHEL7 (Note
that they worked fine on CentOS7, this is RHEL specific.) Fixing
that was first blocked on fixing RHEL7 provisioning as it had some
infra issues.

Now that RHEL 7.2 provisioning is working again, we need to update
these test framework scripts so that upgrade testing on RHEL7 works.

  - when setting the release in the subscription-manager, we need
    to dynamically set the release based on the distro under test,
    and not just have it hard-coded to 6Server.
  - the cluster setup script had a couple hard-coded references
    to rhel 6 that needed updating to be dynamic.

Change-Id: I9d36c0afbbaea7ef6c96405f3e66c2d2670d58d4
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/22503
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>