HYD-6493 Improve Crypto Based on TLS config Doc
Our crypto is currently weak (and weaker still in past releases).
We are currently allowing DH+3DES which should be blacklisted.
We have already disabled SSL variants, but we should also disable older
TLS as it seems we can support it.

Intel has a doc on crypto (attached to ticket) with a ranking of
preferred ciphers. We are able
to support the second cipher in the list without issue,
TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256.

By enabling only TLS 1.2 and TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 we
are in good
standing with Intel's stance on crypto

  - Update Nginx config to only support TLS 1.2.
  - Use TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256 as the 
    only allowed cipher.

Change-Id: I98c486b9c5af4b4ba34163623b1a1e59655ffa3c
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/22524
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>