HYD-5585 o2ib for tcp LNET interface configuration in IML
IML currently restricts the LND type that can be selected for a particular
network type the intention being that a user can only select valid LND types.

In the real world this restriction can lead to valid selections not being
possible for example a connection described by ip as ethernet may actually
be able to support o2ib through RoCE.

This patch updates the Ethernet to LND type supported table to make all
LND types be supported on all Network types. It also moves the table into
settings.py so that a customer could override, extend or modify in some
other way the possibilities.

 - Add o2ib to the tcp entry for the lnd types
 - Move the tables to settings.py

Change-Id: I160440fbc6e890245378ac870af71d44c878505b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22587
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>