HYD-5585 o2ib for tcp LNET interface configuration in IML
IML currently restricts the LND type that can be selected for a particular
network type the intention being that a user can only select valid LND types.

In 2.4 the restriction is that ethernet must be tcp and infiniband must be
o2ib.

In the real world this restriction can lead to valid selections not being
possible for example a connection described by ip as ethernet may actually
be able to support o2ib through RoCE.

This patch enables the mapping to be overridden with new hardcoded values
using the localsettings file.

 - Create a mapping table for network type to lnd type
 - Place the table in settings.py
 - Add the actual changes required from HYD-4961 as the ordered dict issue
   appears to have surfaced on the test systems, presumably due to an upgrade
   in some packages

Change-Id: I160440fbc6e890245378ac870af71d44c878505b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22589
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>